requirements.txt
graphviz
numpy
pandas
pyyaml
scipy
click
pytest
pytest-mock
sphinx
sphinx-autodoc-typehints
sphinx-rtd-theme
